requirements.txt
tensorflow
flask
pillow
numpy
Pillow
scipy
# pip install -r requirements.txt